% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cle.R
\name{cle_ind}
\alias{cle_ind}
\title{Generate data frame for independent mutations model}
\usage{
cle_ind(barge)
}
\arguments{
\item{barge}{List of parameters and data generated using parameter_barge()}
}
\description{
Generate data frame for independent mutations model
}
