% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cle.R
\name{cle_mig}
\alias{cle_mig}
\title{Generate data frame for migration from a source population model}
\usage{
cle_mig(barge)
}
\arguments{
\item{barge}{List of parameters and data generated using parameter_barge()}
}
\description{
Generate data frame for migration from a source population model
}
