% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cle.R
\name{cle_multi}
\alias{cle_multi}
\title{Generate data frame for multiple modes}
\usage{
cle_multi(barge)
}
\arguments{
\item{barge}{List of parameters and data generated using parameter_barge()}
}
\description{
Generate data frame for multiple modes
}
