% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cle.R
\name{cle_neutral}
\alias{cle_neutral}
\title{Generate data frame for neutral model}
\usage{
cle_neutral(barge)
}
\arguments{
\item{barge}{List of parameters and data generated using parameter_barge()}
}
\description{
Generate data frame for neutral model
}
