% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cle.R
\name{cle_svsrc}
\alias{cle_svsrc}
\title{Generate data frame for standing variation from a source population model.}
\usage{
cle_svsrc(barge)
}
\arguments{
\item{barge}{List of parameters and data generated using parameter_barge()}
}
\description{
Generate data frame for standing variation from a source population model.
}
