% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_cle.R
\name{mode_cle}
\alias{mode_cle}
\title{Fit composite likelihood model of convergent adaption.}
\usage{
mode_cle(barge, mode)
}
\arguments{
\item{barge}{List of parameters and data generated using \code{\link{parameter_barge}}}

\item{mode}{Specify the mode of convergent adaption. Options are:

"neutral" -- Neutral model. Return same composite likelihood across all sites.

"independent" --  Selected mutation occurred independently in all populations specified in \code{selected_pops}

 "migration" -- Selected mutation occurs in a source population (post-divergence) and migrates to other populations specified in \code{selected_pops} during the course of the sweep.

 "standing" -- Selected mutation is present at low frequency in ancestral population and sweeps in \code{selected_pops} populations after divergence.

 "standing_source" -- Selected mutation is present at low frequency, experiences migration from a source population prior to selection, sweeps in \code{selected_pops} after migration.

 "multi" --  Some mix of the modes above, with the exception of "standing", which is not currently supported for mixed modes. See \code{sets} and \code{modes} arguments for \code{\link{parameter_barge}} for proper input format.}
}
\value{
Regardless of the mode, the function returns a data frame with the same columns. When a column does not pertain to a mode, \code{NA} values are added to ensure the data frames of different modes can be easily joined. The columns are:

selected_sites -- The proposed sites of selection constructed with or provided to \code{\link{parameter_barge}}.

cle -- The composite likelihood for that row's combination of parameters.

locus -- The user specified name of the input locus. Helpful if multiple loci will be combined in later analysis.

sels -- Selection coefficient.

gs -- Allele frequence of the standing variant prior to sweep.

times -- Time in generations the variant is standing in populations before selection occurs and prior to migration from source population.

migs -- Migration rate.

sources -- The population where the new mutation originated.

sel_pops -- Hyphen delimited string of the selected populations.

model -- The mode of convergent adaptation proposed. When multi modes are used, the modes apply to which
}
\description{
Fit composite likelihood model of convergent adaption.
}
