% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_barge.R
\name{update_mode}
\alias{update_mode}
\title{Update an existing parameter barge with a new mode details.}
\usage{
update_mode(barge, sets, modes)
}
\arguments{
\item{barge}{An existing list object made by calling the parameter_barge() function.}

\item{sets}{List of length number of different modes of convergence to be specified vector "modes" where each element in list contains vector of populations with a given single mode of convergence i.e. if populations 2 and 6 share a mode and populations 3 has another, sets = list(c(2,6), 3).}

\item{modes}{Character vector of length sets defining a new set of mixed modes for each set of selected populations ("ind", "sv", and/or "mig"). Other variables will be updated accordingly.}
}
\description{
Update an existing parameter barge with a new mode details.
}
