% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cle.R
\name{cle_sv}
\alias{cle_sv}
\title{Generate data frame for standing variation model}
\usage{
cle_sv(barge, cores = 1)
}
\arguments{
\item{barge}{List of parameters and data generated using parameter_barge()}

\item{cores}{Number of cores to use. Defaults to 1. if More than one are used, furrr::feautre_pmap is used.}
}
\description{
Generate data frame for standing variation model
}
