% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_barge.R
\name{parameter_barge}
\alias{parameter_barge}
\title{Generate and transfer parameters, quantities, and objects used in a variety of downstream steps to Global environment.}
\usage{
parameter_barge(
  allFreqs,
  freqs_notRand,
  selPops,
  positions,
  n_sites,
  sampleSizes,
  numPops,
  numBins,
  sets = NULL,
  modes = NULL,
  sels,
  migs,
  times,
  gs,
  sources,
  Ne,
  rec,
  locus_name = "locus"
)
}
\arguments{
\item{allFreqs}{Matrix of allele frequencies at putatively neutral sites with dimension number of populations x number of sites.}

\item{freqs_notRand}{Matrix of allele frequencies at putatively selected sites with dimension number of populations x number of sites.}

\item{selPops}{Vector of indices for populations that experienced selection.}

\item{positions}{Vector of genomic positions for selected region.}

\item{n_sites}{Integer for the number of sites to propose as the selected site. Must be less than length(positions).}

\item{sampleSizes}{Vector of sample sizes of length number of populations. (i.e. twice the number of diploid individuals sampled in each population).}

\item{numPops}{Number of populations sampled (both selected and non-selected).}

\item{numBins}{the number of bins in which to bin alleles a given distance from the proposed selected sites.}

\item{sets}{List of length number of different modes of convergence to be specified vector "modes" where each element in list contains vector of populations with a given single mode of convergence i.e. if populations 2 and 6 share a mode and populations 3 has another, sets = list(c(2,6), 3). Required for modeling multiple modes. Only required for fitting mixed mode models.}

\item{modes}{Character vector of length sets defining mode for each set of selected populations ("ind", "sv", and/or "mig"). Only required for fitting mixed mode models.}

\item{sels}{Vector of proposed selection coefficients.}

\item{migs}{Vector of proposed migration rates (proportion of individuals from source each generation).}

\item{times}{Vector of proposed time in generations the variant is standing in populations before selection occurs and prior to migration from source population.}

\item{gs}{Vector of proposed frequencies of the standing variant migs: migration rate (proportion of individuals from source each generation). Note: cannot be 0.}

\item{sources}{Vector of proposed source population of the beneficial allele for both migration and standing variant with source models. Note: the source must be a selected population in selPops.}

\item{Ne}{Effective population size (Assumed equal for all populations).}

\item{rec}{Per base recombination rate for putatively selected region.}

\item{locus_name}{String to name the locus. Helpful if multiple loci will be combined in subsequent analyses. Defaults to "locus"}
}
\description{
Generate and transfer parameters, quantities, and objects used in a variety of downstream steps to Global environment.
}
